+++ mars/oscar/backends/ray/tests/test_ray_actor_context.py
@@ -13,10 +13,11 @@
 # limitations under the License.
 
 import asyncio
-import unittest
+
+import pytest
 
 import mars.oscar as mo
-from mars.tests.core import aio_case, require_ray
+from mars.tests.core import require_ray
 from mars.utils import lazy_import
 
 ray = lazy_import('ray', globals=globals())
@@ -102,117 +103,117 @@ class EventActor(mo.Actor):
         return message
 
 
-@aio_case
 @require_ray
-class Test(unittest.TestCase):
-    def setUp(self) -> None:
+@pytest.mark.asyncio
+class Test:
+    def setup_method(self, _):
         ray.init()
 
-    def tearDown(self) -> None:
+    def teardown_method(self, _):
         ray.shutdown()
 
-    async def testSimpleRayActorPool(self):
+    async def test_simple_ray_actor_pool(self):
         actor_ref = await mo.create_actor(DummyActor, 100, address=RAY_TEST_ADDRESS)
-        self.assertEqual(await actor_ref.add(1), 101)
+        assert await actor_ref.add(1) == 101
         await actor_ref.add(1)
 
         res = await actor_ref.get_value()
-        self.assertEqual(res, 102)
+        assert res == 102
 
         ref2 = await actor_ref.get_ref()
-        self.assertEqual(actor_ref.address, ref2.address)
-        self.assertEqual(actor_ref.uid, ref2.uid)
+        assert actor_ref.address == ref2.address
+        assert actor_ref.uid == ref2.uid
 
-        self.assertEqual(await (await mo.actor_ref(
-                uid=actor_ref.uid, address=actor_ref.address)).add(2), 104)
+        assert await (await mo.actor_ref(
+                uid=actor_ref.uid, address=actor_ref.address)).add(2) == 104
 
-    async def testRayPostCreatePreDestroy(self):
+    async def test_ray_post_create_pre_destroy(self):
         actor_ref = await mo.create_actor(EventActor, address=RAY_TEST_ADDRESS)
         await actor_ref.destroy()
 
-    async def testRayCreateActor(self):
+    async def test_ray_create_actor(self):
         actor_ref = await mo.create_actor(DummyActor, 1, address=RAY_TEST_ADDRESS)
         # create actor inside on_receive
         ref = await actor_ref.create(DummyActor, 5, address=RAY_TEST_ADDRESS)
-        self.assertEqual(await ref.add(10), 15)
+        assert await ref.add(10) == 15
         # create actor inside on_receive and send message
         r = await actor_ref.create_send(DummyActor, 5, method='add', method_args=(1,), address=RAY_TEST_ADDRESS)
-        self.assertEqual(r, 6)
+        assert r == 6
 
-    async def testRayCreateActorError(self):
+    async def test_ray_create_actor_error(self):
         ref1 = await mo.create_actor(DummyActor, 1, uid='dummy1', address=RAY_TEST_ADDRESS)
-        with self.assertRaises(mo.ActorAlreadyExist):
+        with pytest.raises(mo.ActorAlreadyExist):
             await mo.create_actor(DummyActor, 1, uid='dummy1', address=RAY_TEST_ADDRESS)
         await mo.destroy_actor(ref1)
 
         # It's hard to get the exception raised from Actor.__init__.
-        # with self.assertRaises(ValueError):
+        # with pytest.raises(ValueError):
         #     await mo.create_actor(DummyActor, -1, address=RAY_TEST_ADDRESS)
         ref1 = await mo.create_actor(DummyActor, 1, address=RAY_TEST_ADDRESS)
-        with self.assertRaises(ValueError):
+        with pytest.raises(ValueError):
             await ref1.create(DummyActor, -2)
 
-    async def testRaySend(self):
+    async def test_ray_send(self):
         ref1 = await mo.create_actor(DummyActor, 1)
         ref2 = await ref1.create(DummyActor, 2)
-        self.assertEqual(await ref1.send(ref2, 'add', 3), 5)
+        assert await ref1.send(ref2, 'add', 3) == 5
 
-    async def testRaySendError(self):
+    async def test_ray_send_error(self):
         ref1 = await mo.create_actor(DummyActor, 1)
-        with self.assertRaises(TypeError):
+        with pytest.raises(TypeError):
             await ref1.add(1.0)
         ref2 = await mo.create_actor(DummyActor, 2)
-        with self.assertRaises(TypeError):
+        with pytest.raises(TypeError):
             await ref1.send(ref2, 'add', 1.0)
-        with self.assertRaises(mo.ActorNotExist):
+        with pytest.raises(mo.ActorNotExist):
             await (await mo.actor_ref('fake_uid')).add(1)
 
-    async def testRayTell(self):
+    async def test_ray_tell(self):
         ref1 = await mo.create_actor(DummyActor, 1, address=RAY_TEST_ADDRESS)
         ref2 = await ref1.create(DummyActor, 2, address=RAY_TEST_ADDRESS)
-        self.assertIsNone(await ref1.tell(ref2, 'add', 3))
-        self.assertEqual(await ref2.get_value(), 5)
+        assert await ref1.tell(ref2, 'add', 3) is None
+        assert await ref2.get_value() == 5
 
         await ref1.tell_delay(ref2, 'add', 4, delay=.5)  # delay 0.5 secs
-        self.assertEqual(await ref2.get_value(), 5)
+        assert await ref2.get_value() == 5
         await asyncio.sleep(0.5)
-        self.assertEqual(await ref2.get_value(), 9)
+        assert await ref2.get_value() == 9
 
         # error needed when illegal uids are passed
-        with self.assertRaises(TypeError):
+        with pytest.raises(TypeError):
             await ref1.tell(await mo.actor_ref(set()), 'add', 3)
 
-    async def testRayDestroyHasActor(self):
+    async def test_ray_destroy_has_actor(self):
         ref1 = await mo.create_actor(DummyActor, 1)
-        self.assertTrue(await mo.has_actor(ref1))
+        assert await mo.has_actor(ref1)
 
         await mo.destroy_actor(ref1)
-        self.assertFalse(await mo.has_actor(ref1))
+        assert not await mo.has_actor(ref1)
 
         # error needed when illegal uids are passed
-        with self.assertRaises(TypeError):
+        with pytest.raises(TypeError):
             await mo.has_actor(await mo.actor_ref(set()))
 
         ref1 = await mo.create_actor(DummyActor, 1)
         await mo.destroy_actor(ref1)
-        self.assertFalse(await mo.has_actor(ref1))
+        assert not await mo.has_actor(ref1)
 
         ref1 = await mo.create_actor(DummyActor, 1)
         ref2 = await ref1.create(DummyActor, 2)
 
-        self.assertTrue(await mo.has_actor(ref2))
+        assert await mo.has_actor(ref2)
 
         await ref1.delete(ref2)
-        self.assertFalse(await ref1.has(ref2))
+        assert not await ref1.has(ref2)
 
-        with self.assertRaises(mo.ActorNotExist):
+        with pytest.raises(mo.ActorNotExist):
             await mo.destroy_actor(await mo.actor_ref('fake_uid'))
 
         ref1 = await mo.create_actor(DummyActor, 1)
-        with self.assertRaises(mo.ActorNotExist):
+        with pytest.raises(mo.ActorNotExist):
             await ref1.delete(mo.actor_ref('fake_uid'))
 
         # test self destroy
         ref1 = await mo.create_actor(DummyActor, 2)
         await ref1.destroy()
-        self.assertFalse(await mo.has_actor(ref1))
+        assert not await mo.has_actor(ref1)
